tests/wpt/tests/html/canvas/offscreen/reset/2d.reset.state.line_dash.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.reset.state.line_dash</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.reset.state.line_dash</h1>
<p class="desc">check that the line dash is reset</p>


<script>
var t = async_test("check that the line dash is reset");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.setLineDash([1, 2]);

  ctx.reset();
  _assert(ctx.getLineDash().length == 0, "ctx.getLineDash().length == 0");
  t.done();

});
</script>
